feat(nvim): add a lazygit shortcut
Ghostty doesn't have the functionality to launch a new tab, start a
process, and close the tab when the process exits. I have this set up
properly in kitty and when using zellij, but not natively as a part of
the ghostty config. This should probably be removed when
ghostty-org/ghostty#2872 gets implemented.